Compile empty groups before anything
